requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
Pillow>=9.5.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
django-filter>=23.2
celery>=5.3.0
redis>=4.5.5
channels>=4.0.0
drf-yasg>=1.21.5  # For API documentation

# Test dependencies
pytest>=7.3.1
pytest-django>=4.5.2
pytest-cov>=4.1.0  # For coverage reporting
factory-boy>=3.3.0  # For test data factories
faker>=18.13.0  # For generating fake data
freezegun>=1.2.2  # For time-dependent tests
